MFB 64-bit fixes
git-svn-id: https://svn.php.net/repository/pecl/intl/trunk@256693 c90b9560-bf6c-de11-be94-00142212c4b1

+++ normalizer/normalizer_normalize.c
@@ -34,7 +34,7 @@ PHP_FUNCTION( normalizer_normalize )
 {
 	char*			input = NULL;
 	// form is optional, defaults to FORM_C
-	int			form = NORMALIZER_DEFAULT;
+	long			form = NORMALIZER_DEFAULT;
 	int			input_len = 0;
 	zval*			options_array;
 		
@@ -172,7 +172,7 @@ PHP_FUNCTION( normalizer_is_normalized )
 {
 	char*	 	input = NULL;
 	// form is optional, defaults to FORM_C
-	int			form = NORMALIZER_DEFAULT;
+	long		form = NORMALIZER_DEFAULT;
 	int		input_len = 0;
 	zval*	 	options_array = NULL;
